+++ src/ManifoldMakie.jl
@@ -1,0 +1,16 @@
+module ManifoldMakie
+
+using GeometryBasics
+using Manifolds
+using Manifolds: Sphere
+using Makie
+import Makie: scatter, scatter!
+
+include("sphere.jl")
+include("geodesics.jl")
+
+export Sphere
+export scatter, scatter!
+export geodesics, geodesics!
+export scattergeodesics, scattergeodesics!
+end # module ManifoldMakie