+++ src/ManifoldsMakie.jl
@@ -1,5 +1,0 @@
-module ManifoldsMakie
-
-greet() = print("Hello World!")
-
-end # module ManifoldsMakie