+++ Changelog.md
@@ -1,6 +1,6 @@
 # Changelog
 
-All notable Changes to the Julia package `ManifoldsMakie.jl` are documented in this file.
+All notable Changes to the Julia package `ManifoldMakie.jl` are documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
@@ -8,3 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.1.0] unreleased
 
 Initial release.
+
+## Added
+
+* a `sphereplot(Manifolds.Sphere(2))` to start a plot with a sphere to put points on.